refactor(content): ディレクトリ構成を整理
