+++ src/content/index.ts
@@ -1,5 +1,5 @@
 import './style.css';
-import { initializeStorage } from './core/storage';
+import { initializeStorage } from './storage';
 import { initializeResizeHandler } from './ui/renderer';
 import { createObserver, observeYouTubeElements } from './observer';
